added buttons for form, changed auto booleans into ints (re import schema)

+++ sql/yeti_scouting_2018_schema.sql
@@ -32,8 +32,8 @@ DROP TABLE IF EXISTS `form_data`;
 CREATE TABLE `form_data` (
   `id` int(11) NOT NULL,
   `auto_check` tinyint(1) NOT NULL DEFAULT '0',
-  `auto_switch` tinyint(1) NOT NULL DEFAULT '0',
-  `auto_scale` tinyint(1) NOT NULL DEFAULT '0',
+  `auto_switch` int(11) NOT NULL DEFAULT '0',
+  `auto_scale` int(11) NOT NULL DEFAULT '0',
   `auto_speed` int(11) NOT NULL DEFAULT '1',
   `bar_climb` tinyint(1) NOT NULL DEFAULT '0',
   `comment` varchar(500) NOT NULL,
